Partial implementation of ITraversibleList and TraversibleList
Saving this branch to come back and mutation test the code in this
state. I found an interesting issue with iterating using a traverser
using the scenario test. I want to see if that test kills any mutants,
and look at whether any of my other tests should have caught this
problem.